// src/api/models/counselor-list-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CounselorListItem
 */
export interface CounselorListItem {
    /**
     * 
     * @type {string}
     * @memberof CounselorListItem
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof CounselorListItem
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof CounselorListItem
     */
    'roleType'?: CounselorListItemRoleTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof CounselorListItem
     */
    'username'?: string;
    /**
     * 
     * @type {string}
     * @memberof CounselorListItem
     */
    'phoneNumber'?: string;
    /**
     * 
     * @type {string}
     * @memberof CounselorListItem
     */
    'registrationDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof CounselorListItem
     */
    'description'?: string;
}

export const CounselorListItemRoleTypeEnum = {
    Admin: 'ROLE_ADMIN',
    User: 'ROLE_USER',
    Assistant: 'ROLE_ASSISTANT',
    None: 'ROLE_NONE'
} as const;

export type CounselorListItemRoleTypeEnum = typeof CounselorListItemRoleTypeEnum[keyof typeof CounselorListItemRoleTypeEnum];