// src/api/models/update-counselor-req.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateCounselorReq
 */
export interface UpdateCounselorReq {
    /**
     * 
     * @type {string}
     * @memberof UpdateCounselorReq
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateCounselorReq
     */
    'phoneNumber'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateCounselorReq
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateCounselorReq
     */
    'roleType'?: UpdateCounselorReqRoleTypeEnum;
}

export const UpdateCounselorReqRoleTypeEnum = {
    Admin: 'ROLE_ADMIN',
    User: 'ROLE_USER',
    Assistant: 'ROLE_ASSISTANT',
    None: 'ROLE_NONE'
} as const;

export type UpdateCounselorReqRoleTypeEnum = typeof UpdateCounselorReqRoleTypeEnum[keyof typeof UpdateCounselorReqRoleTypeEnum];